ci: simplify tests
